:truck: macro, library function into separate namespaces
Apparently ClojureScript macros are written in JVM Clojure in a
separate namespace and get required with a special :require-macros ns
declaration?

At this point, the Chromium dev console recognizes a Thirty_Three
object with lots of function attributes, so I'm hopeful that all this
can actually be made to work.